# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Amazon Prime Video API – Spécification OpenAPI 3.0
  description: |
    Spécification complète simulant l'API d'Amazon Prime Video.

    ## Domaines couverts
    - **Auth** — Connexion, tokens JWT, 2FA
    - **Profils** — Gestion multi-profils (comme Netflix)
    - **Catalogue** — Films, séries, épisodes, documentaires
    - **Recherche** — Recherche textuelle, par genre, par acteur
    - **🔍 Recherche par image (IA)** — Identifier un film/acteur depuis une photo
    - **Lecteur** — Lecture, reprise, qualité vidéo, sous-titres
    - **Watchlist** — Liste à regarder plus tard
    - **Historique** — Contenus vus et reprises automatiques
    - **Évaluations & Avis** — Notes et commentaires
    - **Téléchargements** — Visionnage hors ligne
    - **Abonnement** — Prime Video, chaînes, add-ons
    - **Paiement** — Transactions, factures
    - **Recommandations** — Personnalisées, tendances, similaires
    - **Notifications** — Nouvelles sorties, rappels
    - **Contrôle parental** — Restrictions par profil
    - **📺 Publicités IA** — Génération automatique de bandes-annonces et spots publicitaires par IA
    - **Live & Events** — Streaming en direct, sport, événements
    - **Langues & Accessibilité** — Sous-titres, audiodescription, multi-langue

    ## Authentification
    Utiliser `POST /auth/login` pour obtenir un `accessToken` JWT (durée 1h).
    À inclure dans toutes les requêtes protégées : `Authorization: Bearer {accessToken}`

...
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via `POST /auth/login`. À inclure dans le header : `Authorization: Bearer {accessToken}`. Durée de validité : 3600 secondes (1 heure).
  schemas:
    Error:
      type: object
      description: |
        Réponse d'erreur standard. `code` est machine-readable (constante), `message` est lisible par un humain, `details` liste les erreurs de validation champ par champ.
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: Le contenu demandé est introuvable.
        details:
          type: array
          items:
            type: string
...
    PaginatedMeta:
      type: object
      description: Wrapper contenant les métadonnées de pagination d'une réponse de liste.
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Duration:
      type: object
      description: Durée d'un contenu vidéo exprimée en heures, minutes et secondes.
      properties:
        hours:
          type: integer
          example: 2
        minutes:
          type: integer
          example: 18
        seconds:
          type: integer
          example: 45
        totalSeconds:
          type: integer
          description: Durée totale en secondes (utile pour les barres de progression)
          example: 8325
    ImageAsset:
      type: object
      description: Asset image (affiche, miniature, bannière) avec ses dimensions.
      properties:
        url:
          type: string
          format: uri
          example: https://cdn.primevideo.com/images/tt0111161_poster.jpg
        width:
          type: integer
          example: 400
        height:
          type: integer
          example: 600
        type:
          type: string
          enum:
            - POSTER
            - THUMBNAIL
            - BANNER
            - HERO
            - LOGO
            - BACKDROP
...
    Person:
      type: object
      description: Personne du monde du cinéma/TV (acteur, réalisateur, scénariste...).
      properties:
        id:
          type: string
          example: person_tom_hanks
        name:
          type: string
          example: Tom Hanks
        role:
          type: string
          description: Rôle dans le contenu (nom du personnage pour les acteurs)
          example: Forrest Gump
        job:
          type: string
          enum:
            - ACTOR
            - DIRECTOR
            - WRITER
            - PRODUCER
            - COMPOSER
            - CINEMATOGRAPHER
            - OTHER
          example: ACTOR
        photoUrl:
          type: string
          format: uri
          nullable: true
    ContentRating:
      type: object
      description: Classification et avertissements de contenu selon le pays.
      properties:
        rating:
          type: string
          description: Classification officielle (ex. CSA pour la France)
          example: '12'
        advisories:
          type: array
          description: Avertissements de contenu (violence, langage, sexualité...)
          items:
            type: string
          example:
            - Violence
            - Langage vulgaire
        country:
          type: string
          example: FR
    ContentSummary:
      type: object
      description: |
        Version allégée d'un contenu vidéo, utilisée dans les listes, carousels et résultats de recherche. Pour le détail complet, utiliser `GET /catalog/{contentId}`.
      properties:
        id:
          type: string
          example: tt0111161
        type:
          type: string
          enum:
            - MOVIE
            - SERIES
            - DOCUMENTARY
            - SHORT
            - SPECIAL
            - LIVE
            - MINI_SERIES
          example: MOVIE
        title:
          type: string
          example: Les Évadés
        originalTitle:
          type: string
          example: The Shawshank Redemption
        year:
          type: integer
          example: 1994
        duration:
          $ref: '#/components/schemas/Duration'
        rating:
          type: number
          format: double
          description: Note moyenne des utilisateurs (0 à 5)
          example: 4.9
        imdbRating:
          type: number
          format: double
          description: Note IMDb (0 à 10)
          example: 9.3
        poster:
          $ref: '#/components/schemas/ImageAsset'
        thumbnail:
          $ref: '#/components/schemas/ImageAsset'
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        isPrime:
          type: boolean
          description: Inclus dans l'abonnement Prime (sinon location/achat requis)
        isNew:
          type: boolean
          description: Ajouté au catalogue dans les 30 derniers jours
        isTrending:
          type: boolean
        contentRating:
          type: string
          example: '12'
        isDownloadable:
          type: boolean
        hasHDR:
          type: boolean
        has4K:
          type: boolean
        hasDolbyAtmos:
          type: boolean
        availableUntil:
          type: string
          format: date
          nullable: true
          description: Date de fin de disponibilité (null = disponible indéfiniment)
    Content:
      type: object
      description: |
        Fiche complète d'un contenu vidéo (film, série, documentaire...). Inclut le casting complet, les saisons/épisodes pour les séries, les options audio/sous-titres disponibles et les liens de lecture.
      allOf:
        - $ref: '#/components/schemas/ContentSummary'
        - type: object
          properties:
            synopsis:
              type: string
              example: Deux hommes emprisonnés à tort se lient d'amitié sur de nombreuses années...
            tagline:
              type: string
              example: Fear can hold you prisoner. Hope can set you free.
            cast:
              type: array
              items:
                $ref: '#/components/schemas/Person'
            directors:
              type: array
              items:
                $ref: '#/components/schemas/Person'
            writers:
              type: array
              items:
                $ref: '#/components/schemas/Person'
            producers:
              type: array
              items:
                $ref: '#/components/schemas/Person'
            studio:
              type: string
              example: Columbia Pictures
            country:
              type: string
              example: US
            languages:
              type: array
              description: Langues audio disponibles
              items:
                type: string
              example:
                - fr
                - en
                - es
                - de
            subtitleLanguages:
              type: array
              description: Langues de sous-titres disponibles
              items:
                type: string
              example:
                - fr
                - en
                - es
                - de
                - it
                - pt
            images:
              type: object
              properties:
                poster:
                  $ref: '#/components/schemas/ImageAsset'
                hero:
                  $ref: '#/components/schemas/ImageAsset'
                banner:
                  $ref: '#/components/schemas/ImageAsset'
                backdrop:
                  $ref: '#/components/schemas/ImageAsset'
                logo:
                  $ref: '#/components/schemas/ImageAsset'
            trailerUrl:
              type: string
              format: uri
              nullable: true
              description: URL de la bande-annonce officielle
            contentRatings:
              type: array
              items:
                $ref: '#/components/schemas/ContentRating'
            videoQualities:
              type: array
              description: Qualités vidéo disponibles pour ce contenu
              items:
                type: string
                enum:
                  - SD
                  - HD
                  - FULL_HD
                  - UHD_4K
            hasAudioDescription:
              type: boolean
              description: Audiodescription disponible (accessibilité malvoyants)
            hasSignLanguage:
              type: boolean
              description: Langue des signes disponible
            seasons:
              type: array
              description: Saisons disponibles (pour les séries uniquement)
              items:
                $ref: '#/components/schemas/SeasonSummary'
            totalEpisodes:
              type: integer
              nullable: true
              description: Nombre total d'épisodes (séries uniquement)
            relatedContent:
              type: array
              description: Contenus similaires recommandés
              items:
                $ref: '#/components/schemas/ContentSummary'
            awards:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: Oscar du Meilleur Film
                  year:
                    type: integer
                    example: 1995
                  won:
                    type: boolean
            pricing:
              $ref: '#/components/schemas/ContentPricing'
            addedAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    ContentList:
      type: object
      description: Liste paginée de contenus (version allégée).
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContentSummary'
        meta:
          $ref: '#/components/schemas/PaginatedMeta'
    SeasonSummary:
      type: object
      description: Résumé d'une saison d'une série (pour les listes).
      properties:
        id:
          type: string
          example: season_001
        number:
          type: integer
          example: 1
        title:
          type: string
          nullable: true
          example: Saison 1
        year:
          type: integer
          example: 2023
        episodeCount:
          type: integer
          example: 8
        thumbnail:
          $ref: '#/components/schemas/ImageAsset'
...
  responses:
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Erreur de validation — les données envoyées sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Numéro de page (commence à 1)
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Nombre d'éléments par page (max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    ContentIdParam:
      name: contentId
      in: path
      required: true
      description: Identifiant unique du contenu (IMDB ID ou ID interne)
      schema:
        type: string
        example: tt0111161
...
    AcceptLanguageParam:
      name: Accept-Language
      in: header
      description: |
        Langue souhaitée pour les contenus localisés (titres, synopsis, interface). Format BCP 47. Si la langue demandée n'est pas disponible, retour en anglais (en-US).
      schema:
        type: string
        example: fr-FR
        default: fr-FR
      required: false
...
paths:
...
  /catalog:
    get:
      tags:
        - Catalog
      summary: Parcourir le catalogue
      operationId: browseCatalog
      description: |
        Liste les contenus du catalogue avec filtres avancés. Retourne des `ContentSummary` (version allégée) pour des performances optimales. Utilisez `GET /catalog/{contentId}` pour le détail complet d'un contenu.
        **Filtres disponibles :** - `type` : MOVIE, SERIES, DOCUMENTARY, SHORT, SPECIAL, LIVE, MINI_SERIES - `genre` : Action, Comédie, Drame, Thriller, Science-fiction... - `year` : Année de sortie - `isPrime` : Uniquement les contenus inclus dans Prime - `hasHDR` / `has4K` : Filtrer par qualité vidéo - `minRating` / `maxRating` : Filtrer par note - `language` : Langue audio disponible
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/AcceptLanguageParam'
        - name: type
          in: query
          schema:
            type: string
            enum:
              - MOVIE
              - SERIES
              - DOCUMENTARY
              - SHORT
              - SPECIAL
              - LIVE
              - MINI_SERIES
        - name: genre
          in: query
          description: ID de genre (séparé par virgule pour plusieurs genres)
          schema:
            type: string
            example: genre_action,genre_thriller
        - name: year
          in: query
          schema:
            type: integer
            example: 2024
        - name: minYear
          in: query
          schema:
            type: integer
            example: 2020
        - name: maxYear
          in: query
          schema:
            type: integer
        - name: isPrime
          in: query
          schema:
            type: boolean
        - name: hasHDR
          in: query
          schema:
            type: boolean
        - name: has4K
          in: query
          schema:
            type: boolean
        - name: hasDolbyAtmos
          in: query
          schema:
            type: boolean
        - name: minRating
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 5
        - name: language
          in: query
          description: Code de langue audio disponible
          schema:
            type: string
            example: fr
        - name: country
          in: query
          description: Pays de production
          schema:
            type: string
            example: US
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - relevance
              - newest
              - oldest
              - rating
              - imdb_rating
              - title_asc
              - title_desc
            default: relevance
        - name: isNew
          in: query
          description: Uniquement les contenus ajoutés dans les 30 derniers jours
          schema:
            type: boolean
        - name: isTrending
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Liste de contenus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentList'
        '400':
          $ref: '#/components/responses/ValidationError'
  /catalog/{contentId}:
    get:
      tags:
        - Catalog
      summary: Récupérer le détail d'un contenu
      operationId: getContent
      description: |
        Retourne la fiche complète d'un contenu (film, série, documentaire...).
        **Contenu de la réponse :** - Titre, synopsis, année, durée, classification - Casting complet (acteurs, réalisateurs, scénaristes) - Galerie d'images (affiche, bannière, logo, backdrop) - URL de bande-annonce - Langues audio et sous-titres disponibles - Qualités vidéo (SD, HD, 4K HDR, Dolby Atmos) - Saisons et épisodes (pour les séries) - Tarification (inclus Prime / location / achat) - Prix et nominations aux récompenses - Contenus similaires recommandés
      security: []
      parameters:
        - $ref: '#/components/parameters/ContentIdParam'
        - $ref: '#/components/parameters/AcceptLanguageParam'
      responses:
        '200':
          description: Détail du contenu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '404':
          $ref: '#/components/responses/NotFound'
...